chore: organize folders structure
