+++ src/main.ts
@@ -1,6 +1,6 @@
 import { NestFactory } from '@nestjs/core';
 import { NestExpressApplication } from '@nestjs/platform-express';
-import { AppModule } from './modules/app.module';
+import { AppModule } from './app/app.module';
 
 const port = process.env.PORT || 4000;
