+++ src/routes/album/album.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { AlbumController } from 'src/controllers/album.controller';
-import { AlbumService } from 'src/providers/album.service';
+import { AlbumController } from 'src/routes/album/album.controller';
+import { AlbumService } from 'src/routes/album/album.service';
 
 @Module({
   imports: [],
