+++ src/routes/album/album.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
-import { albums } from '../db/db';
-import { Album } from '../types/types';
+import { albums } from '../../database/db';
+import { Album } from '../../types/types';
 
 @Injectable()
 export class AlbumService {
