+++ src/routes/artist/artist.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { ArtistController } from 'src/controllers/artist.controller';
-import { ArtistService } from 'src/providers/artist.service';
+import { ArtistController } from 'src/routes/artist/artist.controller';
+import { ArtistService } from 'src/routes/artist/artist.service';
 
 @Module({
   imports: [],
