+++ src/routes/artist/artist.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
-import { artists } from '../db/db';
-import { Artist } from '../types/types';
+import { artists } from '../../database/db';
+import { Artist } from '../../types/types';
 
 @Injectable()
 export class ArtistService {
