+++ src/routes/user/user.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { UserController } from '../controllers/user.controller';
-import { UserService } from '../providers/user.service';
+import { UserController } from './user.controller';
+import { UserService } from './user.service';
 
 @Module({
   imports: [],
