+++ src/routes/user/user.service.ts
@@ -10,8 +10,8 @@ import {
 import * as bcrypt from 'bcrypt';
 import { v4 as uuidv4 } from 'uuid';
 import { validate as uuidValidate } from 'uuid';
-import { users } from '../db/db';
-import { CreateUserDto, UpdatePasswordDto, User } from '../types/types';
+import { users } from '../../database/db';
+import { CreateUserDto, UpdatePasswordDto, User } from '../../types/types';
 
 @Injectable()
 export class UserService {
